requirements.txt
azure-cognitiveservices-vision-computervision==0.7.0
azure-common==1.1.26
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
msrest==0.6.19
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
Pillow==8.0.1
PyMySQL==0.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
Werkzeug==1.0.1
